Documented search handlers.
This commit was primarily focused on documenting the search handlers.
However, I did end up cleaning the workspace search handler a bit.
I did so by moving the event handler in a separate file - `search_workspace_event_helper.js`.
This is exactly the same as yesterday's commit to the toolbox search
handler.
