+++ core/search.js
@@ -1,26 +1,12 @@
 /**
- * @license
- * Visual Blocks Editor
+ * @fileoverview Search handler for blocks inside Blockly.
+ * This is an abstract class that stores blocks and keywords in a search trie.
+ * It should be inherited and extended by *specific* search handlers.
  *
- * Copyright 2012 Google Inc.
- * https://developers.google.com/blockly/
+ * Provides the basic data structure and methods (for initializing, running a search, etc).
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @fileoverview Search handler for Blockly. Gets added to
- * a workspace and handles all searches in that workspace.
+ * Additionally, provides some static functions that are to be
+ * used when building the initial trie (e.g. parsing static keywords).
  *
  * @author ivan@shaperobotics.com
  */
@@ -34,7 +20,7 @@ goog.require('Blockly.Trie');
 /**
  * Initializes the search handler.
  *
- * @param {!Blockly.Workspace} workspace The workspace which the search will work in.
+ * @param {!Blockly.Workspace} workspace The Blockly workspace which the search will work in (will prolly be Blockly.mainWorkspace).
  */
 Blockly.Search = function (workspace) {
   this.workspace_ = workspace;
@@ -42,9 +28,11 @@ Blockly.Search = function (workspace) {
   // Initializes a trie. The trie is the way blocks are stored for quick search results.
   this.blockTrie_ = new Blockly.Trie();
 
-  // Initialize a key-value dictionary which holds the block-keyword connection in reverse to the trie. This is used for easier de-registering of blocks in onBlockRemoved
+  // Initialize a key-value dictionary which holds the block-keyword connection in reverse to the trie.
+  // This is used for easier de-registering of blocks in onBlockRemoved
   this.blocksAdded_ = {};
 
+  // An array that will hold the results immediately after a search.
   this.finalResults_ = [];
 };
 
@@ -56,7 +44,7 @@ Blockly.Search = function (workspace) {
  *
  * @param {!String} key A single keyword
  * @param {!String} value The associated block. Will either be XML
- * (if the Search handler works on the Toolbox) or block type ID (if the Search handler works on a Workspace)
+ * (if the Search handler works on the Toolbox) or block ID (if the Search handler works on a Workspace)
  */
 Blockly.Search.prototype.addToTrie = function (key, value) {
   // Add the keyword to the trie, if it doesn't exist
@@ -77,13 +65,21 @@ Blockly.Search.prototype.addToTrie = function (key, value) {
  * (if the Search handler works on the Toolbox) or the unique block ID (if the Search handler works on a Workspace)
  */
 Blockly.Search.prototype.removeFromTrie = function (key, value) {
+  // Do nothing if the word doesn't exist in the trie
   if (!this.blockTrie_.containsKey(key)) {
     return;
   }
 
+  // Get all of the blocks associated with the word
   var listToClean = this.blockTrie_.get(key);
+
+  // NOTE: if you change this logic, be very careful with list references in JS. I had the problem of removing everything BUT the block before.
+  // If the specified block is in the list...
   if (listToClean.indexOf(value) !== -1) {
+    // Remove the block
     listToClean.pop(value);
+
+    // And put the list back in the trie
     this.blockTrie_.set(key, listToClean);
   }
 };
@@ -92,7 +88,7 @@ Blockly.Search.prototype.removeFromTrie = function (key, value) {
  * Adds a new block to the Search handler's trie.
  *
  * @param {!String} type The type ID of the block, e.g. "fable_play_sound".
- * @param {!String} val The block itself. Similar to addToTrie, will either be XML or the unique block ID.
+ * @param {!String} val The block itself. Similar to `addToTrie`, will either be XML or the unique block ID.
  */
 Blockly.Search.prototype.onBlockAdded = function (type, val) {
   try {
@@ -135,6 +131,7 @@ Blockly.Search.prototype.onBlockRemoved = function (type, val) {
   if (this.blocksAdded_[val]) {
     var keys = this.blocksAdded_[val];
 
+    // Remove each individual keyword <-> block combination
     for (var j = 0; j < keys.length; j++) {
       this.removeFromTrie(keys[j], val);
     }
@@ -142,6 +139,17 @@ Blockly.Search.prototype.onBlockRemoved = function (type, val) {
     // Reset the list of keywords for this block
     delete this.blocksAdded_[val];
   }
+};
+
+/**
+ * Clears the entire trie and reinitializes it.
+ *
+ * Useful when all blocks get permanently removed (e.g. when changing from Simple to Advanced mode).
+ */
+Blockly.Search.prototype.clearAll = function () {
+  delete this.blockTrie_;
+
+  this.blockTrie_ = new Blockly.Trie();
 };
 
 /**
@@ -170,8 +178,11 @@ Blockly.Search.prototype.decodeXmlBlocks = function (xmlBlock) {
 
   allBlocks.push([topMostBlockType, id]);
 
+  var xmlChild;
+
   // Go through the children of the topmost block
-  for (var i = 0, xmlChild; xmlChild = xmlBlock.childNodes[i]; i++) {
+  for (var i = 0; i < xmlBlock.childNodes.length; i++) {
+    xmlChild = xmlBlock.childNodes[i];
     if (xmlChild.nodeType === 3) {
       // Ignore any text at the <block> level.  It's all whitespace anyway.
       continue;
@@ -181,8 +192,11 @@ Blockly.Search.prototype.decodeXmlBlocks = function (xmlBlock) {
     var childBlockElement = null;
     var childShadowElement = null;
 
-    // Get the current child. Either a block or a shadow block
-    for (var j = 0, grandchild; grandchild = xmlChild.childNodes[j]; j++) {
+    var grandchild;
+    for (var j = 0; j < xmlChild.childNodes.length; j++) {
+      // Get the current child. Either a block or a shadow block
+      grandchild = xmlChild.childNodes[j];
+
       if (grandchild.nodeType === 1) {
         if (grandchild.nodeName.toLowerCase() === 'block') {
           childBlockElement = /** @type {!Element} */ (grandchild);
@@ -207,8 +221,10 @@ Blockly.Search.prototype.decodeXmlBlocks = function (xmlBlock) {
         if (childBlockElement) {
           var blockChild = this.decodeXmlBlocks(childBlockElement);
 
-          for (var k = 0, grandchild; grandchild = blockChild[k]; k++) {
-            allBlocks.push(grandchild);
+          var grandgrandchild;
+          for (var k = 0; k < blockChild.length; k++) {
+            grandgrandchild = blockChild[k];
+            allBlocks.push(grandgrandchild);
           }
         }
         break;
@@ -225,18 +241,24 @@ Blockly.Search.prototype.decodeXmlBlocks = function (xmlBlock) {
  *
  * @param {!String} term Single word that will be used to search for blocks.
  *
- * @return {!Array<String>} A list of all block IDs (or XML, see onBlockAdded) that match the keyword.
+ * @return {!Array<String>} A list of all results (block XMLs or block IDs, see `onBlockAdded`) that match the keyword.
  */
 Blockly.Search.prototype.blocksMatchingSearchTerm = function (term) {
   if (!this.blockTrie_) {
     return [];
   }
 
-  var keys = this.blockTrie_.getKeys(term.toLowerCase());
+  // Initialize the holder of final results
   var blocks = [];
 
+  // Get all separate words
+  var keys = this.blockTrie_.getKeys(term.toLowerCase());
+
+  // Iterate through the keywords
   for (var i = 0; i < keys.length; i++) {
     var key = keys[i];
+
+    // Get all blocks associated with a single keyword
     var blocksForKey = this.blockTrie_.get(key);
 
     for (var j = 0; j < blocksForKey.length; j++) {
@@ -252,9 +274,12 @@ Blockly.Search.prototype.blocksMatchingSearchTerm = function (term) {
  * the entire string the user entered. Will be split into single words and
  * the results for each single word will be intersected.
  *
+ * Results from separate keywords will be intersected. This means a block will only be part of the results
+ * if it matches ALL keywords provided.
+ *
  * @param {!Array<String>} terms Entire contents of the user's search. Must be in a list.
  *
- * @return {!Array<String>} A list of all block IDs (or XML, see onBlockAdded) that match the user's search.
+ * @return {!Array<String>} A list of all results (block XMLs or block IDs, see `onBlockAdded`) that match the keyword.
  */
 Blockly.Search.prototype.blocksMatchingSearchTerms = function (terms) {
   var intersectingMatches = null;
@@ -279,6 +304,15 @@ Blockly.Search.prototype.blocksMatchingSearchTerms = function (terms) {
   return intersectingMatches.getValues();
 };
 
+/**
+ * Executes a single search based on a search term.
+ * The term gets cleaned up (separate words split, all lowercase, etc)
+ * and the search is conducted.
+ *
+ * @param {String} inputValue the string which represents the search terms.
+ *
+ * @returns {Array<Object>} an array of results that match the search term.
+ */
 Blockly.Search.prototype.runSearch = function (inputValue) {
   // Prepare the contents of the search by trimming, lowercasing and splitting by whitespace
   var searchTerms = inputValue.trim().toLowerCase().split(/\s+/);
@@ -289,31 +323,34 @@ Blockly.Search.prototype.runSearch = function (inputValue) {
   });
 
   // Temporary list for results
-  var matchingBlockIds = [];
+  var matchingResults = [];
 
+  // Run the actual search (if there are any terms to search with)
   if (searchTerms.length > 0) {
-    matchingBlockIds = this.blocksMatchingSearchTerms(searchTerms);
+    matchingResults = this.blocksMatchingSearchTerms(searchTerms);
   }
 
-  return matchingBlockIds;
+  // Return the results
+  return matchingResults;
 };
 
 /**
- * Clears the Search handler's trie and reinitializes it.
+ * Execute this when the user "ends" a search by closing it.
  *
- * Useful when all blocks get permanently removed (e.g. when changing from Simple to Advanced mode).
+ * Resets the array which holds results.
  */
-Blockly.Search.prototype.clearAll = function () {
-  delete this.blockTrie_;
-
-  this.blockTrie_ = new Blockly.Trie();
-};
-
 Blockly.Search.prototype.onCloseSearch = function () {
   this.finalResults_ = [];
 };
 
-// Filters out empty spaces. Stolen from Closure's goog.array.filter so we can remove it completely.
+/** =================================================
+ *  =============== STATIC FUNCTIONS ================
+ *  =================================================
+ */
+
+/**
+ * Filters out empty spaces. Stolen from Closure's goog.array.filter so we can remove it completely.
+ */
 Blockly.Search.filter = function (arr, f, optObj) {
   var l = arr.length; // must be fixed during loop... see docs
   var res = [];
@@ -348,16 +385,22 @@ Blockly.Search.preprocessSearchKeywords = function (blockType, staticKeywords, e
     Blockly.Search.initStaticKeywords(staticKeywords, Blockly.Blocks[blockType].StaticSearchKeywords);
   }
 
+  // Check if the block has any "extra" keywords associated with the toolbox search only
   if (!extraToolboxKeywords) {
     return;
   }
 
+  // Check if the "extra" toolbox keywords have been initialized before, if not, do it now.
   if (!Blockly.Blocks[blockType].StaticToolboxSearchKeywords || Blockly.Blocks[blockType].StaticToolboxSearchKeywords.length === 0) {
     Blockly.Blocks[blockType].StaticToolboxSearchKeywords = [];
     Blockly.Search.initStaticKeywords(extraToolboxKeywords, Blockly.Blocks[blockType].StaticToolboxSearchKeywords);
   }
 };
 
+/**
+ * Internal function for preprocessing blocks' associated keywords.
+ * See `Blockly.Search.preprocessSearchKeywords` for mroe details
+ */
 Blockly.Search.initStaticKeywords = function (keywordList, arrayToUse) {
   // Go through the provided list of words/sentences
   for (let i = 0; i < keywordList.length; i++) {
@@ -385,17 +428,26 @@ Blockly.Search.initStaticKeywords = function (keywordList, arrayToUse) {
   }
 };
 
+/**
+ * Static function for "cleaning" keyword <-> block associations
+ * whenever the user changes the app's language.
+ * This is done so the trie can then be rebuilt with the new language strings.
+ */
 Blockly.Search.onLanguageChange = function () {
+  // Get all blocks in the app
   const keys = Object.keys(Blockly.Blocks);
 
   if (keys.length > 0) {
+    // Go through the blocks
     for (let i = 0; i < keys.length; i++) {
       const singleKey = keys[i];
 
+      // Clean out any static keywords
       if (Blockly.Blocks[singleKey].StaticSearchKeywords) {
         Blockly.Blocks[singleKey].StaticSearchKeywords.length = 0;
       }
 
+      // Clean out any "extra" toolbox keywords
       if (Blockly.Blocks[singleKey].StaticToolboxSearchKeywords) {
         Blockly.Blocks[singleKey].StaticToolboxSearchKeywords.length = 0;
       }
